# src/main/resources/static/api-docs.json
{"openapi":"3.0.1","info":{"title":"Employee API","description":"API documentation for managing students","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/employees/{id}":{"get":{"tags":["employee-controller"],"operationId":"getEmployeeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDTO"}}}}}},"put":{"tags":["employee-controller"],"operationId":"updateEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployessRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDTO"}}}}}},"delete":{"tags":["employee-controller"],"operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/employees":{"get":{"tags":["employee-controller"],"operationId":"getAllEmployees","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponseDTO"}}}}}}},"post":{"tags":["employee-controller"],"operationId":"createEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployessRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDTO"}}}}}}},"/employees/{id}/profile-image":{"get":{"tags":["employee-controller"],"operationId":"getImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}},"post":{"tags":["employee-controller"],"operationId":"upload","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/hii":{"get":{"tags":["hello"],"operationId":"hii","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v2/employees":{"get":{"tags":["employee-uri-versioning"],"operationId":"v2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/employees":{"get":{"tags":["employee-uri-versioning"],"operationId":"v1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/employees":{"get":{"tags":["employee-media-type-versioning"],"operationId":"v2_1_1_1_1_1_1","parameters":[{"name":"api-version","in":"header","schema":{"type":"string","enum":["2","1"]}}],"responses":{"200":{"description":"OK","content":{"application/vnd.employee-v2+json":{"schema":{"type":"object"}},"application/vnd.employee-v1+json":{"schema":{"type":"object"}},"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"EmployessRequestDTO":{"required":["department","email","name"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"},"department":{"type":"string"}}},"EmployeeResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"getname":{"type":"string"},"email":{"type":"string"},"creatdAt":{"type":"string","format":"date-time"},"department":{"type":"string"}}}}}}